% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_count.R
\name{char_count}
\alias{char_count}
\title{Characters count}
\usage{
char_count(x)
}
\arguments{
\item{x}{- character vector}
}
\value{
data.frame with two columns: charater and number of its occurencies.
}
\description{
Characters count
}
\examples{
char_count("character")
char_count(words.eng)
}

