% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_word.R
\name{find_word}
\alias{find_word}
\title{Find words that fit the chosen parameters}
\usage{
find_word(model = "*", allow = letters, ban = character(0),
  type = "usual", words = words.eng)
}
\description{
Find words that fit the chosen parameters
}
\examples{
find_word("w..d")
find_word(allow = "thing", type = "scrabble")
find_word(allow = "thing", type = "anagram")
}

