% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelToRegex.R
\name{model_to_regex}
\alias{model_to_regex}
\title{Build a regular expression to fit chosen parameters}
\usage{
model_to_regex(model = "*", allow = letters, ban = character(0),
  type = "usual")
}
\arguments{
\item{model}{- model...}

\item{allow}{- characters allowed to be in a word. Can be listed in a
single string or as vector ...}

\item{ban}{- characters not allowed to be in a word.}

\item{type}{- can be \code{"usual"}, \code{"scrabble"}, or \code{"anagram"}.
match.arg is used ...}
}
\description{
Build a regular expression to fit chosen parameters
}
\examples{
model_to_regex("w..d")
model_to_regex(allow = "thing", type = "scrabble")
model_to_regex(allow = "thing", type = "anagram")
}

