% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_word.R
\name{scrabble}
\alias{anagram}
\alias{scrabble}
\title{Find words that can be constructed from the specified letters}
\usage{
scrabble(allow, model = "*", ban = character(0), words = words.eng)

anagram(allow, model = "*", ban = character(0), words = words.eng)
}
\description{
\code{scrabble} function finds the words that can be constructed
from the specified set of letters. \cr
\code{anagram} function finds the words that are permutations of the specified
set of letters. Usually this set of letters is a word itself.\cr
\code{scrabble} and \code{anagram} are functions built on top of
the \code{\link{find_word}} function with parameter \code{type} set to "scrabble"
or "anagram" respectively and \code{allow} parameter moved to the first place
to simplify the usage (see examples).
}
\seealso{
\code{\link{find_word}}
}

