% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words.eng}
\alias{words.eng}
\title{English words list}
\format{An object of class \code{character} of length 263533.}
\description{
263,533 english words list took from \url{http://norvig.com/ngrams/}
(See \code{word.list} file).
}
\keyword{datasets}

