% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_count.R
\name{char_count}
\alias{char_count}
\title{Characters count}
\usage{
char_count(x)
}
\arguments{
\item{x}{character vector, or a list that can be unlisted to a character
vector.}
}
\value{
data.frame with two columns: \code{char} - character and \code{count}
  - number of it's occurencies.
}
\description{
Calculates character frequencies in a vector.
}
\examples{
char_count("character")
char_count(words.eng)
}

