% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_regex.R
\name{model_to_regex}
\alias{model_to_regex}
\title{Build a regular expression to fit chosen parameters}
\usage{
model_to_regex(model = "*", allow = letters, ban = character(0),
  type = "usual")
}
\arguments{
\item{model}{- pattern that a word should match. Is constructed of \code{.}
possibly followed by {...} repetition quantifier, ...}

\item{allow}{- characters allowed to be in a word. Can be listed in a
single string or in a vector.}

\item{ban}{- characters not allowed to be in a word.}

\item{type}{- can be \code{"usual"}, \code{"scrabble"}, or \code{"anagram"}.
Abbreviated input is allowed: e.g. \code{"u"}, \code{"s"},
 or \code{"a"}. See details. (?)}
}
\description{
description
}
\details{

}
\examples{
model_to_regex("w..d")
model_to_regex("p.{2,3}", ban = "aorp")
model_to_regex(allow = "thing", type = "scrabble")
model_to_regex(allow = "thing", type = "anagram")
}

