% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_word.R
\name{scrabble}
\alias{scrabble}
\title{Find words that can be constructed from the specified letters}
\usage{
scrabble(x, model = "*")
}
\description{
Find words that can be constructed from the specified letters
}
\examples{
scrabble("abll")
}

